Separate LemonCalendarRangeInline
Note breaking change: The selected dates selector is now not conditional on no of months shown
