+++ frontend/src/scenes/plugins/edit/interface-jobs/PluginJobConfiguration.tsx
@@ -10,9 +10,9 @@ import { JobPayloadFieldOptions, JobSpec } from '~/types'
 import { interfaceJobsLogic } from './interfaceJobsLogic'
 import { LemonInput } from '../../../../lib/components/LemonInput/LemonInput'
 import moment from 'moment'
-import { LemonCalendarRange } from '../../../../lib/components/LemonCalendarRange/LemonCalendarRange'
 import { LemonModal } from '../../../../lib/components/LemonModal'
 import { LemonButton } from '../../../../lib/components/LemonButton'
+import { LemonCalendarRangeInline } from '../../../../lib/components/LemonCalendarRange/LemonCalendarRangeInline'
 
 interface PluginJobConfigurationProps {
     jobName: string
@@ -148,6 +148,6 @@ function FieldInput({
                 />
             )
         case 'daterange':
-            return <LemonCalendarRange value={value || null} onChange={onChange} hideHeader autoApply />
+            return <LemonCalendarRangeInline value={value || null} onChange={onChange} />
     }
 }
